Preparing main.rs for ratatui
