+++ Cargo.toml
@@ -10,3 +10,8 @@ thirtyfour = "0.35.0"
 futures = "0.3.31"
 toml = "0.7"
 serde = { version = "1.0.217", features = ["derive"] }
+
+crossterm = "0.28.1"
+ratatui = "0.30.0-alpha.1"
+tui-input = "0.11.1"
+rand = "0.8.5"
